ctrlsw: Align VCU ControlSW with revision 30233
Added features (Codec) :
- fix crashes when the dma allocation failed
- fix freqIDR
- external SEI support (encode)
- Fix command line parsing : hexadecimal value are allowed again

API:
- export less headers in include: We do not export IScheduler.h anymore.
- exe_encoder and exe_decoder do not depend on any internal headers anymore
- add API to add section in the stream buffer from outside lib_encode (external SEI support)
- add decoder AL_Decoder_ForceStop API to terminate decoding regardless of remaining data in
- the circular buffer and/or input buffer(s), only wait for end of command already sent to the MCU

+++ lib_encode/hardwareDriver.c
@@ -53,7 +53,7 @@ void Close(Driver* driver, int fd)
   Rtos_DriverClose((void*)(uintptr_t)fd);
 }
 
-bool PostMessage(Driver* driver, int fd, int messageId, void* data)
+bool PostMessage(Driver* driver, int fd, long unsigned int messageId, void* data)
 {
   (void)driver;
   /* must keep the errno from ioctl */
